refactor: merge app package into cmd
