+++ cmd/entropy/version.go
@@ -1,1 +1,17 @@
 package main
+
+import (
+	"github.com/spf13/cobra"
+
+	v "github.com/odpf/entropy/pkg/version"
+)
+
+func cmdVersion() *cobra.Command {
+	return &cobra.Command{
+		Use:   "version",
+		Short: "Show version information",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return v.Print()
+		},
+	}
+}
